ACTIONS-FIND NAME ON CALI
This script searches a CALI listing for a partial exact match of a
client's name.  It can search first name, last name, or both, and
results will include partial matches.  It has no soundex.  This script
was written for use in Anoka County - some modification will be needed
for other counties to use this script.