See #419
This update makes this script usable for the rest of the Collaborative (by offering counties except Anoka the user the option to type in mod worker name, title and phone to update mod worker info).  The script checks the cali code to identify Anoka users, and for Anoka users, displays a dropdown for mod workers instead of the text fields.   Again, this script does not do all that was initially requested to do.  However, it is a great first step that can be enhanced in the future.

Anoka County recently had a telecommunications upgrade that changed all of our phone numbers, and this fix also updates Anoka County mod worker phone numbers.