Fix bugs
Utilized the write_date function to populate the date for adding the worklist in the script.  Note that I am not able to test this fix because with our system settings, I never encounter this problem.  Also fixed a bug - we had a missing transmit that would effect collaborative script users. With this fix I believe this is ready for other counties to test!!